Move `<excerpt>` component to src/sidebar/components
